feat!: rename package to safeband
Rename the import package functions/ -> safeband/ and the distribution
adaptive-interpretable-ad -> safeband. Updates all imports (library,
tests, example notebooks/scripts, rpc/consumer, pulsar services),
doctests, coverage omit paths, README header/branding, and the lockfile.

BREAKING CHANGE: import path changes from 'functions.*' to 'safeband.*';
distribution name changes from 'adaptive-interpretable-ad' to 'safeband'.
