+++ examples/11_i7_ablations.py
@@ -41,9 +41,9 @@ from river.utils import Rolling
 
 sys.path.insert(1, str(Path(__file__).resolve().parent.parent))
 
-from functions.anomaly import ConditionalGaussianScorer
-from functions.fault_diagnosis import SensorFaultClassifier
-from functions.proba import MultivariateGaussian
+from safeband.anomaly import ConditionalGaussianScorer
+from safeband.fault_diagnosis import SensorFaultClassifier
+from safeband.proba import MultivariateGaussian
 
 logger = logging.getLogger(__name__)
