+++ safeband/typing_extras.py
