+++ tests/test_rpc_server_sink.py
@@ -11,14 +11,14 @@ from streamz import Stream
 sys.path.insert(1, str(Path(__file__).parent.parent))
 
 import rpc_server
-from functions.typing_extras import (
+from rpc_server import RpcOutlierDetector
+from safeband.typing_extras import (
     FileClient,
     KafkaClient,
     MQTTClient,
     NATSClient,
     PulsarClient,
 )
-from rpc_server import RpcOutlierDetector
 
 
 class TestDumpToFile:
